Added more profile functionality Top 3 Most played champions now visible Faster Summoner Searching Ability to edit status and view status

+++ Project Nesja/Objects/Summoner.cs
@@ -3,7 +3,7 @@
     public class Summoner
     {
         public string? PUUID;
-        public int AccountID;
+        public long AccountID;
         public int SummonerID;
         public string? Name;
         public string? InternalName;
